queer-sim-backend/requirements.txt
fastapi
uvicorn
httpx
numpy
python-multipart
yt-dlp
faster-whisper
openai
langdetect
ffmpeg-python
python-dotenv
google-adk
google-genai
nest-asyncio
